+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/ptv/resources/PtvOrganizationServiceResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.ptv.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.ptv.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractOrganizationServiceResourceContainer;
 
 @ApplicationScoped
-public class PtvOrganizationServiceCache extends AbstractOrganizationServiceResourceContainer {
+public class PtvOrganizationServiceResourceContainer extends AbstractOrganizationServiceResourceContainer {
 
   private static final long serialVersionUID = 6100260276077235996L;
